pyproject.toml
[project]
name = "emis-backend"
version = "1.0.0"
description = "Thapathali Campus EMIS Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi==23.1.0",
    "argon2-cffi-bindings==25.1.0",
    "asgiref==3.10.0",
    "attrs==25.4.0",
    "beautifulsoup4==4.12.2",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "cfgv==3.4.0",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "colorama==0.4.6",
    "coverage==7.5.0",
    "cryptography==46.0.3",
    "cssbeautifier==1.15.4",
    "distlib==0.4.0",
    "Django==4.2.2",
    "django-ckeditor==6.6.1",
    "django-cors-headers==4.6.0",
    "django-csp==3.7",
    "django-environ==0.11.2",
    "django-filter==24.3",
    "django-jazzmin==3.0.1",
    "django-js-asset==2.0.0",
    "django-mptt==0.16.0",
    "django-oauth-toolkit==2.2.0",
    "django-redis==5.4.0",
    "djangorestframework==3.15.2",
    "djangorestframework-camel-case==1.4.2",
    "djangorestframework-simplejwt==5.3.1",
    "djlint==1.34.1",
    "drf-nested-forms==1.1.8",
    "drf-spectacular==0.28.0",
    "EditorConfig==0.17.1",
    "filelock==3.20.0",
    "gunicorn==20.1.0",
    "html-tag-names==0.1.2",
    "html-void-elements==0.1.0",
    "identify==2.6.15",
    "idna==3.11",
    "inflection==0.5.1",
    "jsbeautifier==1.15.4",
    "json5==0.9.28",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "jwcrypto==1.5.6",
    "lxml==4.9.3",
    "nodeenv==1.9.1",
    "oauthlib==3.3.1",
    "pathspec==0.12.1",
    "pillow==10.3.0",
    "platformdirs==4.5.0",
    "pre-commit==3.7.0",
    "psycopg2-binary==2.9.10",
    "pycparser==2.23",
    "PyJWT==2.10.1",
    "PyYAML==6.0.3",
    "redis==7.0.1",
    "referencing==0.37.0",
    "regex==2023.12.25",
    "requests==2.32.2",
    "rpds-py==0.28.0",
    "ruff==0.4.2",
    "setuptools==80.9.0",
    "six==1.17.0",
    "soupsieve==2.8",
    "sqlparse==0.5.3",
    "tqdm==4.67.1",
    "typing_extensions==4.15.0",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "virtualenv==20.35.4",
    "whitenoise==6.6.0",
    "xmltodict==0.13.0",
]

# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = [
    "tests.py",
    "test_*.py",
]

# ==== Coverage ====
[tool.coverage.run]
include = ["django_starter/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H030,H031,T002,PLR0913"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

# Exclude a variety of commonly ignored directories.
[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "*/migrations/*.py",
    "staticfiles/*"
]
# Same as Django: https://github.com/cookiecutter/cookiecutter-django/issues/4792.
line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "C90",
  "I",
  "N",
  "UP",
  "YTT",
  # "ANN", # flake8-annotations: we should support this in the future but 100+ errors atm
  "ASYNC",
  "S",
  "BLE",
  "FBT",
  "B",
  "A",
  "COM",
  "C4",
  "DTZ",
  "T10",
  "DJ",
  "EM",
  "EXE",
  "FA",
  'ISC',
  "ICN",
  "G",
  'INP',
  'PIE',
  "T20",
  'PYI',
  'PT',
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SLOT",
  "SIM",
  "TID",
  "TCH",
  "INT",
  # "ARG", # Unused function argument
  "PTH",
  "ERA",
  "PD",
  "PGH",
  "PL",
  "TRY",
  "FLY",
  # "NPY",
  # "AIR",
  "PERF",
  # "FURB",
  # "LOG",
  "RUF"
]
ignore = [
"EXE002",
  "S101", # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "SIM102", # sometimes it's better to nest
  "UP038", # Checks for uses of isinstance/issubclass that take a tuple
          # of types for comparison.
          # Deactivated because it can make the code slow:
          # https://github.com/astral-sh/ruff/issues/7871
"PLR0913",
"F401"
]
# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []
# The fixes in extend-unsafe-fixes will require
# provide the `--unsafe-fixes` flag when fixing.
extend-unsafe-fixes = [
    "UP038"
]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# Sorting the imports
[tool.ruff.lint.isort]
force-single-line = false
case-sensitive = false
force-wrap-aliases = true
combine-as-imports = true
constants = []
